refactor: 과목 반영 결과 계층 의존성 정리
