+++ src/main/java/inu/timetable/dto/SubjectImportApplyResponse.java
@@ -1,7 +1,5 @@
 package inu.timetable.dto;
 
-import inu.timetable.service.TimetableConflictResolutionService;
-
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -15,7 +13,7 @@ public record SubjectImportApplyResponse(
         int modifiedCount,
         int canceledCount,
         SubjectImportImpactResponse previewedImpact,
-        TimetableConflictResolutionService.ReconciliationResult reconciliation,
+        TimetableReconciliationResult reconciliation,
         List<VerificationItem> verification) {
 
     public record VerificationItem(String courseCode, boolean matched, String message) {
